requirements.txt
flask
beautifulsoup4
requests
apscheduler
flask-sqlalchemy